gramvikash_api/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
google-genai==1.65.0
firebase-admin==6.5.0
djangorestframework-simplejwt==5.3.1
boto3==1.34.0
langdetect==1.0.9
Pillow==11.3.0
fastapi==0.115.0
uvicorn==0.30.6
websockets==13.1
edge-tts